Move metadata operations into DotnetInspector.Metadata library (#23)
Make DotnetInspector.Metadata a self-contained query API for .NET assembly
metadata. The CLI becomes a pure rendering layer — other tools can build on
the Metadata library without pulling in CLI concerns.

Phase 1 — Data model + ApiSurfaceExtractor:
- Move ApiSurface/ApiType/ApiMember POCOs to Metadata (no Markout attrs)
- Move ApiSurfaceExtractor to Metadata namespace
- Add AssemblyReader convenience layer (hides PEReader boilerplate)
- Add CliApiSurface wrapper in CLI for --all-types Markout serialization
- Update DiffCommand, FindCommand, PlatformCommand, ImplementsCommand

Phase 2 — ExtensionMethodScanner:
- Consolidate extension scanning into Metadata using SignatureDecoder
- Eliminate duplicate ExtensionSignatureTypeProvider from CLI
- Move FindReachableTypes for --reachable flag

Phase 3 — SourceLinkResolver:
- Move to Metadata with new string-based ResolveTypeSource overload
- Add static ExtractRepositoryUrl convenience method
- Remove FindTypeDefinitionHandle from ApiServices (no longer needed)

Phase 4 — AssemblyInspector + AssemblyAuditor:
- Move AssemblyInfo/AssemblyReference/AssemblyReferenceNode to Metadata
- Create AssemblyInspector for PE header, audit, and reference extraction
- Create AssemblyAuditInfo POCO for audit results
- Rewrite AssemblyAuditor to delegate all SRM work to AssemblyInspector

SRM imports reduced from 10 CLI files to 2 (AssemblyCommand, ApiServices),
both retained only for the PDB download pipeline which exposes SRM types.
