+++ src/DotnetInspector.Metadata/DotnetInspector.Metadata.csproj
@@ -14,6 +14,11 @@
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
   </PropertyGroup>
 
+  <ItemGroup>
+    <InternalsVisibleTo Include="dotnet-inspect" />
+    <InternalsVisibleTo Include="dotnet-inspect.Tests" />
+  </ItemGroup>
+
   <!-- System.Reflection.Metadata is part of net10.0 BCL -->
 
 </Project>
