+++ src/dotnet-inspect/AssemblyAudit.cs
@@ -1,4 +1,5 @@
 using System.Text.Json.Serialization;
+using DotnetInspector.Metadata;
 using Markout;
 
 namespace DotnetInspector;
