+++ src/dotnet-inspect/Commands/SamplesCommand.cs
@@ -1,5 +1,6 @@
 using System.Text;
 using DotnetInspector.Inspectors;
+using DotnetInspector.Metadata;
 using DotnetInspector.Options;
 using DotnetInspector.Output;
 using DotnetInspector.Packages;
