+++ src/dotnet-inspect/Inspectors/DocCommentParser.cs
@@ -3,6 +3,7 @@ using System.Buffers;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Xml;
+using DotnetInspector.Metadata;
 
 namespace DotnetInspector.Inspectors;
