+++ src/dotnet-inspect/JsonContext.cs
@@ -1,4 +1,5 @@
 using System.Text.Json.Serialization;
+using DotnetInspector.Metadata;
 
 namespace DotnetInspector;
